chore: clean up privy svm register wallet
