+++ apps/web/src/lib/wallet/namespaces/svm/provider/svm-wallet-provider.tsx
@@ -25,6 +25,7 @@ import type { Wallet } from 'src/lib/wallet/types'
 import { SvmChainId } from 'sushi/svm'
 import type { WalletNamespaceContext } from '../../types'
 import { SvmAdapterId } from '../config'
+import { useRegisterPrivySvmWallet } from './use-register-privy-svm-wallet'
 
 function useInSvmContext(): boolean {
   const client = useConnectorClient()
@@ -60,6 +61,7 @@ export default function SvmWalletProvider({
 
 function _SvmWalletProvider({ children }: { children: React.ReactNode }) {
   const client = useConnectorClient()
+  useRegisterPrivySvmWallet(client)
   const privyEmbeddedWallet = usePrivyEmbeddedWallet('svm')
   const walletInfo = useWalletInfo()
   const {
